tests/wpt/tests/navigation-api/navigate-event/replaceState-inside-back-handler-infinite.optional.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../navigation-methods/return-value/resources/helpers.js"></script>
<script>
promise_test(async t => {
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));
  await navigation.navigate("#push").finished;
  let { reject, promise } = Promise.withResolvers();
  navigation.addEventListener("navigate", () => {
    try {
      history.replaceState(null, "", "#");
    } catch (e) {
      reject(e);
    }
  });
  await Promise.all([
    promise_rejects_dom(t, "SecurityError", promise, "Infinite recursion in replaceState"),
    assertBothRejectDOM(t, navigation.back(), "AbortError")]);
}, "infinite replaceState inside a navigate event for navigation.back()");
</script>
